Refresh deployments when changing w3 attribute on a package
